moving daq_scan into dedicated folder
