+++ packages/pymodaq/src/pymodaq/extensions/__init__.py
@@ -8,10 +8,9 @@ from .utils import get_extensions
 from .pid.utils import get_models
 
 from .console import QtConsole
-from .daq_scan import DAQScan
+from pymodaq.extensions.scan.daq_scan import DAQScan
 from .daq_logger.daq_logger import DAQ_Logger
 from .pid.pid_controller import DAQ_PID
-from .h5browser import H5Browser  #backcompat but should be loaded from pymodaq_gui!
 
 from .bayesian.bayesian_optimization import BayesianOptimization
 from .bayesian.utils import OptimizerModelDefault
