+++ packages/pymodaq/src/pymodaq/utils/h5modules/module_saving.py
@@ -6,7 +6,7 @@ Created the 23/11/2022
 """
 from __future__ import annotations
 
-from typing import Union, List, Dict, Tuple, TYPE_CHECKING, Iterable
+from typing import Union, List, Tuple, TYPE_CHECKING, Iterable
 import xml.etree.ElementTree as ET
 
 
@@ -15,20 +15,18 @@ from pymodaq_utils.logger import set_logger, get_module_name
 from pymodaq_utils.enums import BaseEnum
 from pymodaq_utils.abstract import ABCMeta, abstract_attribute, abstractmethod
 from pymodaq_utils.utils import capitalize
-from pymodaq_data.data import Axis, DataDim, DataWithAxes, DataToExport, DataDistribution
+from pymodaq_data.data import Axis, DataToExport, DataDistribution
 from pymodaq_data.h5modules.saving import H5SaverLowLevel
-from pymodaq_data.h5modules.backends import GROUP, CARRAY, Node, GroupType
+from pymodaq_data.h5modules.backends import GROUP, Node
 from pymodaq_data.h5modules.data_saving import (
-    DataToExportSaver, AxisSaverLoader, DataToExportEnlargeableSaver,
+    DataToExportSaver, DataToExportEnlargeableSaver,
     DataToExportTimedSaver, DataToExportExtendedSaver)
 from pymodaq_gui.parameter import ioxml
 
 if TYPE_CHECKING:
-    from pymodaq.extensions.daq_scan import DAQScan
+    from pymodaq.extensions.scan.daq_scan import DAQScan
     from pymodaq.control_modules.daq_viewer import DAQ_Viewer
     from pymodaq.control_modules.daq_move import DAQ_Move
-    from pymodaq.extensions.daq_logger.h5logging import H5Logger
-
 
 logger = set_logger(get_module_name(__file__))
